+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- New student preferences design
+- Preferences for half days
+- More preferences 
+- Ability to save preferences
+
 ### Changed
 
 ### TBD